REFACTO T0+2

+++ ComputerDataBase/src/com/excilys/model/java/Company.java
@@ -1,14 +1,14 @@
 package com.excilys.model.java;
 //Model company
 public class Company {
-	private int id;
+	private long id;
 	private String name;
 
 
 	public Company(){}
 
 
-	public Company(int id, String name) {
+	public Company(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
@@ -18,11 +18,11 @@ public class Company {
 		return "Company [id=" + id + ", name=" + name + "]";
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
