+++ ComputerDataBase/src/com/excilys/persistence/java/CompanyDAOInterface.java
@@ -5,6 +5,6 @@ import java.util.List;
 import com.excilys.model.java.Company;
 
 public interface CompanyDAOInterface {
-	public abstract Company find(int id);
+	public abstract Company find(long id);
 	public abstract List<Company>  findAll(); 
 }
